Add an option to git scanner which enable entropy calculation for matched results
The entropy is calculated for each word of a matched line considering only the
word with the maximum entropy. Based on this entropy value, the findings with the
entropy value under a defined threshold are filter out.

This allow us to remove many false positives which are for instance only
paths in a secrets manager.

Signed-off-by: Cosmin Cojocar <cosmin.cojocar@gmx.ch>